+++ src/consts.rs
@@ -1,6 +1,6 @@
 use egui::Color32;
 
-pub const CELL_SIZE: usize = 5;
-pub const CELL_BORDER_SIZE: usize = 1;
+pub const CELL_SIZE: f32 = 5.;
+pub const CELL_BORDER_SIZE: f32 = 1.;
 pub const BOARD_SIZE: usize = 200;
 pub const BACKGROUND_COLOR: Color32 = Color32::GRAY;
